refactor: lift HTTP byte helpers to stackchan-net::http_parse
The previous two PRs (#159, #160) added `cfg(test)` blocks to the
firmware crate for `parse_content_length`, `parse_bearer_token`,
`ct_eq`, and the small byte helpers — but the firmware crate
compiles only for `xtensa-esp32s3-none-elf` and `just check`
explicitly excludes it, so those tests never ran. Lifting the
helpers into a new `stackchan-net::http_parse` module gives them
real host-side coverage.

The cut-over surfaced one real bug: the orphaned test asserted
that `Content-Length: +5` would be rejected, but
`usize::from_str("+5")` actually returns `Ok(5)`. RFC 9110 §8.6
specifies `Content-Length = 1*DIGIT`, so the parser now enforces
digits-only before delegating to `from_str`. Two PRs of CI never
caught this because the test never ran — exactly the failure mode
this refactor is designed to prevent.

`parse_content_length` returns `Result<usize, ParseError>` instead
of `Result<usize, HttpError>` so `stackchan-net` doesn't pull in
firmware-specific error types. The single firmware caller maps
`ParseError → HttpError::Malformed` at the call site. The
firmware's `cfg(test)` module is trimmed back to the
`emotion_str_round_trips_through_parser` test (still orphaned —
`Emotion` is firmware-domain — but documented as such).
